feat: refactor model provider quota handling and improve type safety
- Moved ModelProviderQuotaGetPaid enum to a new file for better organization.
- Updated imports across components to use the new type definition.
- Enhanced type safety in utility functions by specifying more precise types.
- Simplified provider configuration by deriving provider list from a shared constant.
